+++ core/src/test/java/cucumber/runtime/StepDefinitionMatchTest.java
@@ -3,7 +3,6 @@ package cucumber.runtime;
 import cucumber.runtime.converters.LocalizedXStreams;
 import gherkin.I18n;
 import gherkin.formatter.Argument;
-import gherkin.formatter.Reporter;
 import gherkin.formatter.model.DocString;
 import gherkin.formatter.model.Step;
 import org.junit.Test;
@@ -14,6 +13,7 @@ import java.util.List;
 
 import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -25,7 +25,6 @@ public class StepDefinitionMatchTest {
     @Test
     public void converts_numbers() throws Throwable {
         StepDefinition stepDefinition = mock(StepDefinition.class);
-        Reporter reporter = mock(Reporter.class);
         List<ParameterType> parameterTypes = asList(new ParameterType(Integer.TYPE, null));
         when(stepDefinition.getParameterTypes()).thenReturn(parameterTypes);
 
@@ -41,7 +40,6 @@ public class StepDefinitionMatchTest {
     @Test
     public void can_have_doc_string_as_only_argument() throws Throwable {
         StepDefinition stepDefinition = mock(StepDefinition.class);
-        Reporter reporter = mock(Reporter.class);
         List<ParameterType> parameterTypes = asList(new ParameterType(String.class, null));
         when(stepDefinition.getParameterTypes()).thenReturn(parameterTypes);
 
@@ -58,7 +56,6 @@ public class StepDefinitionMatchTest {
     @Test
     public void can_have_doc_string_as_last_argument_among_many() throws Throwable {
         StepDefinition stepDefinition = mock(StepDefinition.class);
-        Reporter reporter = mock(Reporter.class);
         List<ParameterType> parameterTypes = asList(new ParameterType(Integer.TYPE, null), new ParameterType(String.class, null));
         when(stepDefinition.getParameterTypes()).thenReturn(parameterTypes);
 
@@ -73,15 +70,17 @@ public class StepDefinitionMatchTest {
     }
 
     @Test
-    public void retrieve_step_name() {
-        String theName = "name";
+    public void throws_arity_mismatch_exception() throws Throwable {
+        Step step = new Step(null, "Given ", "I have 4 cukes in my belly", 1, null, null);
 
-        Step step = mock(Step.class);
-        when(step.getName()).thenReturn(theName);
-
-        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(null, mock(StepDefinition.class), null, step, null);
-
-        String stepName = stepDefinitionMatch.getStepName();
-        assertEquals(theName, stepName);
+        StepDefinition stepDefinition = new StubStepDefinition(new Object(), Object.class.getMethod("toString"), "some pattern");
+        StepDefinitionMatch stepDefinitionMatch = new StepDefinitionMatch(asList(new Argument(7, "3")), stepDefinition, null, step, new LocalizedXStreams(getClass().getClassLoader()));
+        try {
+            stepDefinitionMatch.runStep(new I18n("en"));
+            fail();
+        } catch(CucumberException expected) {
+            assertEquals("Arity mismatch: Step Definition 'toString' with pattern /some pattern/ is declared with 0 parameters. However, the gherkin step matched 1 arguments [3]. \n" +
+                    "Step: Given I have 4 cukes in my belly", expected.getMessage());
+        }
     }
 }