+++ core/src/main/java/cucumber/table/TableConverter.java
@@ -3,6 +3,7 @@ package cucumber.table;
 import com.thoughtworks.xstream.XStream;
 import com.thoughtworks.xstream.converters.Converter;
 import com.thoughtworks.xstream.converters.SingleValueConverter;
+import com.thoughtworks.xstream.converters.reflection.AbstractReflectionConverter;
 import com.thoughtworks.xstream.io.HierarchicalStreamReader;
 import cucumber.runtime.CucumberException;
 import cucumber.table.xstream.DataTableWriter;
@@ -43,7 +44,11 @@ public class TableConverter {
         } else {
             reader = new ListOfObjectReader(itemType, convertedAttributeNames(dataTable), dataTable.cells(1));
         }
-        return (List) xStream.unmarshal(reader);
+        try {
+            return (List) xStream.unmarshal(reader);
+        } catch (AbstractReflectionConverter.UnknownFieldException e) {
+            throw new CucumberException(e.getShortMessage());
+        }
     }
 
     /**